por ahora

+++ app/Http/Controllers/API/CentroController.php
@@ -17,11 +17,17 @@ class CentroController extends Controller
      *
      * @return \Illuminate\Http\Response
      */
-    public function index()
+
+    public function indexOD()
     {
         // return CentroResource::collection(Centro::paginate(10));
         $response = Http::get('https://datosabiertos.regiondemurcia.es/catalogo/api/action//datastore_search?resource_id=52dd8435-46aa-495e-bd2b-703263e576e7&limit=5');
         return response()->json(json_decode($response));
+    }
+
+    public function index()
+    {
+        return CentroResource::collection(Centro::paginate(10));
     }
 
     /**
@@ -32,6 +38,8 @@ class CentroController extends Controller
      */
     public function store(Request $request)
     {
+
+        $this->authorize('restore', Centro::class);
         $centro = json_decode($request->getContent(), true);
 
         $centro = Centro::create($centro);
@@ -63,9 +71,11 @@ class CentroController extends Controller
     public function update(Request $request, Centro $centro)
     {
 
-        if (! Gate::allows('update-centro', $centro)) {
+        $this->authorize('update', $centro);
+
+        /*if (! Gate::allows('update-centro', $centro)) {
             abort(403);
-        }
+        }*/
 
         $centroData = json_decode($request->getContent(), true);
         $centro->update($centroData);
@@ -79,8 +89,14 @@ class CentroController extends Controller
      * @param  \App\Models\Centro  $centro
      * @return \Illuminate\Http\Response
      */
-    public function destroy(Centro $centro)
+    public function destroy(Request $request, Centro $centro)
     {
+
+        /*if ($request->user()->cannot('delete', $centro)) {
+            abort(403, 'NO eres coordinador de este centro');
+        }*/
+
+        $this->authorize('delete', $centro);
         $centro->delete();
     }
 }
